Created test for channel message factory

+++ test/com/sundays/chat/server/channel/dummy/DummyChannelDataIO.java
@@ -49,7 +49,7 @@ public class DummyChannelDataIO implements ChannelDataManager {
 	}
 
 	@Override
-	public void changeRank(int channelID, int userID, int rankID) {
+	public void changeRank(int channelID, int userID, byte rankID) {
 		calls.add(new CallEvent("changeRank", channelID, userID, rankID));
 	}
